backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.30
pymysql>=1.1.1
cryptography>=42.0.8
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.3
pydantic>=2.7.4
pydantic-settings>=2.3.4
scikit-learn>=1.5.0
pandas>=2.2.2
numpy>=1.26.4
python-multipart>=0.0.9
openpyxl>=3.1.4
reportlab>=4.2.0
joblib>=1.4.2
